production_requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0

# Otimizações de produção
django-cors-headers==4.3.1
django-health-check==3.16.7
redis==5.0.1
django-redis==5.4.0
